services/workers/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
google-cloud-documentai==2.26.0
google-cloud-firestore==2.14.0
vertexai==1.38.1
pydantic==2.6.1
